[Frontend] Unify plugin permissions row styling with install dialog
The plugin detail Permissions section used its own terser row layout
(icon + label + raw detail), which diverged from the install
CapabilitiesWarning dialog (icon + label + description + monospace
detail). Extract a shared CapabilityRow component and capability
definitions so both surfaces render identically, and have
PluginPermissions reuse the permission subset (network, filesystem,
ffmpeg, shell).
