+++ app/components/plugins/CapabilitiesWarning.tsx
@@ -1,14 +1,5 @@
-import {
-  ArrowRightLeft,
-  FileOutput,
-  FileSearch,
-  FolderOpen,
-  Globe,
-  Search,
-  Terminal,
-  Video,
-  type LucideIcon,
-} from "lucide-react";
+import { CapabilityRow } from "./CapabilityRow";
+import { CAPABILITY_DEFS, filterDefsByCapabilities } from "./capabilityRows";
 
 import { Button } from "@/components/ui/button";
 import {
@@ -19,10 +10,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import type {
-  AvailablePlugin,
-  PluginCapabilities,
-} from "@/hooks/queries/plugins";
+import type { AvailablePlugin } from "@/hooks/queries/plugins";
 
 interface CapabilitiesWarningProps {
   isPending: boolean;
@@ -30,127 +18,6 @@ interface CapabilitiesWarningProps {
   onOpenChange: (open: boolean) => void;
   open: boolean;
   plugin: AvailablePlugin | null;
-}
-
-interface CapabilityRowProps {
-  description: string;
-  detail?: string;
-  icon: LucideIcon;
-  label: string;
-}
-
-const CapabilityRow = ({
-  description,
-  detail,
-  icon: Icon,
-  label,
-}: CapabilityRowProps) => (
-  <div className="flex items-start gap-3 rounded-md border border-border p-3">
-    <Icon className="mt-0.5 h-4 w-4 shrink-0 text-muted-foreground" />
-    <div>
-      <p className="text-sm font-medium">{label}</p>
-      <p className="text-xs text-muted-foreground">{description}</p>
-      {detail && (
-        <p className="mt-1 font-mono text-xs text-muted-foreground/70">
-          {detail}
-        </p>
-      )}
-    </div>
-  </div>
-);
-
-interface CapabilityDef {
-  key: keyof PluginCapabilities;
-  icon: LucideIcon;
-  label: string;
-  description: string;
-  detail: (cap: PluginCapabilities) => string | undefined;
-}
-
-const CAPABILITY_DEFS: CapabilityDef[] = [
-  {
-    key: "metadataEnricher",
-    icon: Search,
-    label: "Metadata Enrichment",
-    description: "Searches external sources for book metadata",
-    detail: (cap) =>
-      cap.metadataEnricher?.fileTypes?.length
-        ? cap.metadataEnricher.fileTypes.join(", ")
-        : undefined,
-  },
-  {
-    key: "inputConverter",
-    icon: ArrowRightLeft,
-    label: "Format Conversion",
-    description: "Converts files between formats",
-    detail: (cap) =>
-      cap.inputConverter?.sourceTypes?.length && cap.inputConverter?.targetType
-        ? `${cap.inputConverter.sourceTypes.join(", ")} \u2192 ${cap.inputConverter.targetType}`
-        : undefined,
-  },
-  {
-    key: "fileParser",
-    icon: FileSearch,
-    label: "File Parsing",
-    description: "Extracts metadata from files",
-    detail: (cap) =>
-      cap.fileParser?.types?.length
-        ? cap.fileParser.types.join(", ")
-        : undefined,
-  },
-  {
-    key: "outputGenerator",
-    icon: FileOutput,
-    label: "Output Generation",
-    description: "Generates files in additional formats",
-    detail: (cap) =>
-      cap.outputGenerator?.sourceTypes?.length && cap.outputGenerator?.name
-        ? `${cap.outputGenerator.sourceTypes.join(", ")} \u2192 ${cap.outputGenerator.name}`
-        : undefined,
-  },
-  {
-    key: "httpAccess",
-    icon: Globe,
-    label: "Network Access",
-    description: "May make network requests to external services",
-    detail: (cap) =>
-      cap.httpAccess?.domains?.length
-        ? cap.httpAccess.domains.join(", ")
-        : undefined,
-  },
-  {
-    key: "fileAccess",
-    icon: FolderOpen,
-    label: "File System Access",
-    description: "Can access files beyond its sandboxed plugin directory",
-    detail: (cap) => {
-      if (cap.fileAccess?.level === "readwrite") return "read/write";
-      if (cap.fileAccess?.level === "read") return "read-only";
-      return undefined;
-    },
-  },
-  {
-    key: "ffmpegAccess",
-    icon: Video,
-    label: "FFmpeg Execution",
-    description: "May invoke FFmpeg for media processing",
-    detail: () => undefined,
-  },
-  {
-    key: "shellAccess",
-    icon: Terminal,
-    label: "Shell Command Execution",
-    description: "May execute shell commands on your system",
-    detail: (cap) =>
-      cap.shellAccess?.commands?.length
-        ? cap.shellAccess.commands.join(", ")
-        : undefined,
-  },
-];
-
-function getCapabilityRows(capabilities: PluginCapabilities | undefined) {
-  if (!capabilities) return [];
-  return CAPABILITY_DEFS.filter((def) => capabilities[def.key] != null);
 }
 
 export const CapabilitiesWarning = ({
@@ -165,7 +32,10 @@ export const CapabilitiesWarning = ({
   const latestVersion =
     plugin.versions.find((v) => v.compatible) ?? plugin.versions[0];
 
-  const rows = getCapabilityRows(latestVersion?.capabilities);
+  const rows = filterDefsByCapabilities(
+    CAPABILITY_DEFS,
+    latestVersion?.capabilities,
+  );
 
   return (
     <Dialog onOpenChange={onOpenChange} open={open}>
